Moves incoming activitypub json to module
