+++ fedireads/activitypub/__init__.py
@@ -1,6 +1,7 @@
 ''' bring activitypub functions into the namespace '''
 from .actor import get_actor
-from .collection import get_outbox, get_outbox_page, get_add, get_remove
+from .collection import get_outbox, get_outbox_page, get_add, get_remove, \
+        get_following, get_followers
 from .create import get_create
 from .follow import get_follow_request, get_accept
 from .status import get_review, get_status
