+++ fedireads/activitypub/collection.py
@@ -33,15 +33,65 @@ def get_outbox_page(user, page_id, statuses, max_id, min_id):
     return page
 
 
-def get_ordered_collection(path, size):
+def get_followers(user, page, follow_queryset):
+    ''' list of people who follow a user '''
+    id_slug = '%s/followers' % user.actor
+    return get_follow_info(id_slug, page, follow_queryset)
+
+
+def get_following(user, page, follow_queryset):
+    ''' list of people who follow a user '''
+    id_slug = '%s/following' % user.actor
+    return get_follow_info(id_slug, page, follow_queryset)
+
+
+def get_follow_info(id_slug, page, follow_queryset):
+    ''' a list of followers or following '''
+    if page:
+        return get_follow_page(follow_queryset, id_slug, page)
+    count = follow_queryset.count()
+    return {
+        '@context': 'https://www.w3.org/ns/activitystreams',
+        'id': id_slug,
+        'type': 'OrderedCollection',
+        'totalItems': count,
+        'first': '%s?page=1' % id_slug,
+    }
+
+
+# TODO: generalize these pagination functions
+def get_follow_page(user_list, id_slug, page):
+    ''' format a list of followers/following '''
+    page = int(page)
+    page_length = 10
+    start = (page - 1) * page_length
+    end = start + page_length
+    follower_page = user_list.all()[start:end]
+    data = {
+        '@context': 'https://www.w3.org/ns/activitystreams',
+        'id': '%s?page=%d' % (id_slug, page),
+        'type': 'OrderedCollectionPage',
+        'totalItems': user_list.count(),
+        'partOf': id_slug,
+        'orderedItems': [u.actor for u in follower_page],
+    }
+    if end <= user_list.count():
+        # there are still more pages
+        data['next'] = '%s?page=%d' % (id_slug, page + 1)
+    if start > 0:
+        data['prev'] = '%s?page=%d' % (id_slug, page - 1)
+    return data
+
+
+def get_ordered_collection(id_slug, size):
     ''' create an ordered collection '''
     return {
         '@context': 'https://www.w3.org/ns/activitystreams',
-        'id': path,
+        'id': id_slug,
         'type': 'OrderedCollection',
         'totalItems': size,
-        'first': '%s?page=true' % path,
-        'last': '%s?min_id=0&page=true' % path
+        'first': '%s?page=true' % id_slug,
+        'last': '%s?min_id=0&page=true' % id_slug
     }
