+++ fedireads/incoming.py
@@ -141,7 +141,8 @@ def get_followers(request, username):
 
     user = models.User.objects.get(localname=username)
     followers = user.followers
-    return format_follow_info(user, request.GET.get('page'), followers)
+    page = request.GET.get('page')
+    return JsonResponse(activitypub.get_followers(user, page, followers))
 
 
 @csrf_exempt
@@ -152,45 +153,8 @@ def get_following(request, username):
 
     user = models.User.objects.get(localname=username)
     following = models.User.objects.filter(followers=user)
-    return format_follow_info(user, request.GET.get('page'), following)
-
-
-def format_follow_info(user, page, follow_queryset):
-    ''' create the activitypub json for followers/following '''
-    id_slug = '%s/following' % user.actor
-    if page:
-        return JsonResponse(get_follow_page(follow_queryset, id_slug, page))
-    count = follow_queryset.count()
-    return JsonResponse({
-        '@context': 'https://www.w3.org/ns/activitystreams',
-        'id': id_slug,
-        'type': 'OrderedCollection',
-        'totalItems': count,
-        'first': '%s?page=1' % id_slug,
-    })
-
-
-def get_follow_page(user_list, id_slug, page):
-    ''' format a list of followers/following '''
-    page = int(page)
-    page_length = 10
-    start = (page - 1) * page_length
-    end = start + page_length
-    follower_page = user_list.all()[start:end]
-    data = {
-        '@context': 'https://www.w3.org/ns/activitystreams',
-        'id': '%s?page=%d' % (id_slug, page),
-        'type': 'OrderedCollectionPage',
-        'totalItems': user_list.count(),
-        'partOf': id_slug,
-        'orderedItems': [u.actor for u in follower_page],
-    }
-    if end <= user_list.count():
-        # there are still more pages
-        data['next'] = '%s?page=%d' % (id_slug, page + 1)
-    if start > 0:
-        data['prev'] = '%s?page=%d' % (id_slug, page - 1)
-    return data
+    page = request.GET.get('page')
+    return JsonResponse(activitypub.get_following(user, page, following))
 
 
 def handle_incoming_follow(activity):
